refactor(types): DereferenceMemberMatcher + комбинируем высокоуровневые методы
По замечанию майнтейнера: unknownMemberReceiverAt/isUnknownGlobalAt делали всё в
обход уже имеющихся высокоуровневых методов, заново их комбинируя — упростил
до прямой комбинации публичных методов.

  isUnknownGlobalAt(doc, pos):
    return findMembersAt(doc, pos).isEmpty()
      && referenceResolver.findReference(doc.getUri(), pos).isEmpty();

  unknownMemberReceiverAt(doc, pos):
    if (!findMembersAt(doc, pos).isEmpty()) return Optional.empty();
    return receiverTypesAt(doc, pos).filter(allConcrete);

Контракт «позиция указывает на нужный вид идентификатора» теперь у вызывающих
(visitGlobalMethodCall / visitAccessProperty / visitMethodCall в
UnknownMemberDiagnostic) — это явно отмечено в javadoc публичных методов.

По S1448 (TypeService 36 методов > 35) — вынесен @Component
DereferenceMemberMatcher, забравший 9 единиц: matchMembers, MatchContext record,
collectCanonicalMembers, collectLocalFieldMembers, countMeaningfulArgs,
inferArgTypes, findDereferenceForTerminal, rightMatchesTerminal плюс полный
конвейер matchAt. У matcher'а есть публичный receiverTypesAt — на нём
выстроен новый высокоуровневый TypeService.receiverTypesAt.

TypeServiceDelegationTest: добавлен @Mock DereferenceMemberMatcher в
конструкторе TypeService.
